requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pyyaml==6.0.2
kubernetes==30.1.0
langchain-google-genai==2.0.9
langchain-core==0.3.37